+++ src/organisations/organisation-versions.service.ts
@@ -79,13 +79,14 @@ export class OrganisationVersionsService {
     );
   }
 
-  async allDraftOrLive(): Promise<Organisation[]> {
+  async allWithLatestVersion(): Promise<Organisation[]> {
     const versions = await this.versionsWithJoins()
       .distinctOn(['organisationVersion.organisation'])
       .where('organisationVersion.status IN(:...status)', {
         status: [
           OrganisationVersionStatus.Live,
           OrganisationVersionStatus.Draft,
+          OrganisationVersionStatus.Archived,
         ],
       })
       .orderBy(
@@ -138,6 +139,37 @@ export class OrganisationVersionsService {
       }
 
       version.status = OrganisationVersionStatus.Live;
+      await this.repository.save(version);
+
+      await queryRunner.commitTransaction();
+    } catch (err) {
+      await queryRunner.rollbackTransaction();
+    } finally {
+      await queryRunner.release();
+    }
+
+    return version;
+  }
+
+  async archive(version: OrganisationVersion): Promise<OrganisationVersion> {
+    const queryRunner = this.connection.createQueryRunner();
+    const organisation = version.organisation;
+
+    const liveVersion = await this.repository.findOne({
+      organisation,
+      status: OrganisationVersionStatus.Live,
+    });
+
+    await queryRunner.connect();
+    await queryRunner.startTransaction();
+
+    try {
+      if (liveVersion) {
+        liveVersion.status = OrganisationVersionStatus.Draft;
+        await this.repository.save(liveVersion);
+      }
+
+      version.status = OrganisationVersionStatus.Archived;
       await this.repository.save(version);
 
       await queryRunner.commitTransaction();